feat: initial implementation of get and set (#3)
Adds an initial implementation of the `MomentoRedisReactiveClient` and
integration tests:
- Adds the class `MomentoRedisReactiveClient` which implements the
Lettuce Reactive Redis command set interface. This is a client side
proxy that takes a Momento cache client and cache name as dependencies
in the constructor. Implements the `get` and `set` commands in the first
cut.
- Adds a narrow interface `MomentoRedisReactiveCommands` with the
specific commands we have implemented. This allows a developer the
confidence to only use the commands we have authored.
- Adds a `RedisCodec` wrapper `RedisCodecByteArrayConverter` for
Momento: Lettuce is generic over the key and value types, allowing a
developer to use whatever they want, provided they pass a `RedisCodec`
in the constructor to marshal/unmarshal the data to `ByteBuffer`.
Because Momento takes `byte[]` and not `ByteBuffer` we provide this
wrapper for internal convenience.
- Adds a `MomentoLettuceExceptionMapper` to convert Momento exceptions
back to Lettuce exceptions. The relevant exception type here is the
timeout exception.
- Adds integration tests that can run vs Momento or vs Redis. The
`BaseTestClass` sets up the reactive client, which may be backed by
either Lettuce or our Momento implementation. To toggle Redis vs
Momento, set `REDIS=1` in the environment for Redis. The single test
class `ScalarCommandTest` shows how we have a single set of integration
tests that can run.
- Updates the Makefile to run tests vs Redis, vs Momento, or vs both.
- Updates GitHub CI to run vs Redis and vs Momento

In a future PR we will update the other commands to throw an exception
if not implemented.