feat: add pexpire command (#5)
Implements the `pexpire` command. These command comes with a variety of
options, such as to update the TTL:
- only if no TTL has been set
- only if a TTL has been set
- only when the provided TTL is greater than the existing one
- only when the provided TTL is less than the existing one

Currently the Java SDK only supports an unconditional ttl update, so we
implement that one first and error out on the other options. To
summarize, if a user provides the option for...
- ... `only if no TTL has been set`: we throw an exception since Momento
keys always have a TTL
- ... `only if a TTL has been set`: this is a don't-care since Momento
keys always have a TTL
- ... `only when the provided TTL is greater than the existing one`:
error since the Java SDK doesn't have `increaseTtl`
- `" less than "`: " ... `decreaseTtl`

When no options are provided, this means perform an unconditional update
on the TTL, which we do support. We should communicate with the users to
understand if they are using these options, or if an unconditional
update suffices.